Add files via upload
The file pbnumbers.txt contains the winning lottery numbers selected weekly between 
February 3, 2010 and May 11, 2016. (The file contains 654 sets of winning numbers).
Each line in the file contains the set of six winning numbers that we selected on a given 
date. The numbers are separated by a space and the last number in each line is the 
complementary number for that day. For example, the first line in the file shows the 
numbers for February 3 2010, which were 17, 22, 36, 37, 52 and the complementary 
number is 24.